* Split safeArea
