+++ themes/default/config/character-select.json5
@@ -1,8 +1,8 @@
 {
     theme: {
         "character-select": {
-            rows: 2,
-            columns: 3,
+            rows: 1,
+            columns: 1,
             ease: "Cubic.easeInOut",
             duration: 500,
             align: "left",
